requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
PyMySQL
python-dotenv
requests